Moved project to top-level directory
- It's single module maven project, move it to the top-level
  directory.
- Changed "core" to "googlegroupconnecor" in package path
  since it's not really core app, it's own module.
